+++ 10-file-system/08-my-rm-dir.js
@@ -1,4 +1,10 @@
 #!/usr/bin/node
 const fs = require('fs'),
       dir = process.argv[2];
-fs.rmdirSync(dir);
+if(fs.existsSync(dir)) {
+    if(fs.statSync(dir).isDirectory())  fs.rmdirSync(dir);
+
+}else{
+    console.error('%s not exist!', dir);
+      process.exit(1);
+}
